+++ openeo.d.ts
@@ -5,7 +5,7 @@ import { ProcessRegistry } from '@openeo/js-commons';
 import { Readable } from 'stream';
 import axios from 'axios';
 
-declare namespace OpenEO {
+declare namespace Client {
     /**
      * The base class for authentication providers such as Basic and OpenID Connect.
      *
@@ -356,12 +356,6 @@ declare namespace OpenEO {
          * @throws {Error}
          */
         protected validate(): void;
-        /**
-         * Initializes the class.
-         *
-         * @protected
-         */
-        protected init(): void;
         /**
          * Returns the capabilities response as a JSON serializable representation of the data that is API compliant.
          *
@@ -2125,15 +2119,6 @@ declare namespace OpenEO {
          */
         protected processes: ProcessRegistry;
         /**
-         * Initializes the connection by requesting the capabilities.
-         *
-         * @async
-         * @protected
-         * @returns {Promise<Capabilities>} Capabilities
-         * @throws {Error}
-         */
-        protected init(): Promise<Capabilities>;
-        /**
          * Refresh the cache for processes.
          *
          * @async
@@ -2539,13 +2524,11 @@ declare namespace OpenEO {
          *
          * @async
          * @param {Process} process - A user-defined process.
-         * @param {?string} [plan=null] - The billing plan to use for this computation.
-         * @param {?number} [budget=null] - The maximum budget allowed to spend for this computation.
          * @param {?AbortController} [abortController=null] - An AbortController object that can be used to cancel the processing request.
          * @param {object.<string, *>} [additional={}] - Other parameters to pass for the batch job, e.g. `log_level`.
          * @returns {Promise<SyncResult>} - An object with the data and some metadata.
          */
-        computeResult(process: Process, plan?: string | null, budget?: number | null, abortController?: AbortController | null, additional?: Record<string, any>): Promise<SyncResult>;
+        computeResult(process: Process, abortController?: AbortController | null, additional?: Record<string, any>): Promise<SyncResult>;
         /**
          * Executes a process synchronously and downloads to result the given path.
          *
@@ -2646,23 +2629,16 @@ declare namespace OpenEO {
          */
         getService(id: string): Promise<Service>;
         /**
-         * Get the a link with the given rel type.
+         * Adds additional response details to the array.
+         *
+         * Adds links and federation:missing.
          *
          * @protected
-         * @param {Array.<Link>} links - An array of links.
-         * @param {string|Array.<string>} rel - Relation type(s) to find.
-         * @returns {string | null}
-         * @throws {Error}
+         * @param {Array.<*>} arr
+         * @param {object.<string, *>} response
+         * @returns {ResponseArray}
          */
-        protected _getLinkHref(links: Array<Link>, rel: string | Array<string>): string | null;
-        /**
-         * Makes all links in the list absolute.
-         *
-         * @param {Array.<Link>} links - An array of links.
-         * @param {?string|AxiosResponse} [base=null] - The base url to use for relative links, or an response to derive the url from.
-         * @returns {Array.<Link>}
-         */
-        makeLinksAbsolute(links: Array<Link>, base?: (string | AxiosResponse) | null): Array<Link>;
+        protected _toResponseArray(arr: Array<any>, response: object<string, any>): ResponseArray;
         /**
          * Sends a GET request.
          *
@@ -2782,7 +2758,7 @@ declare namespace OpenEO {
      *
      * @hideconstructor
      */
-    export class OpenEO {
+    export class Client {
         /**
          * Connect to a back-end with version discovery (recommended).
          *
@@ -2819,7 +2795,7 @@ declare namespace OpenEO {
          */
         static clientVersion(): string;
     }
-    export namespace OpenEO {
+    export namespace Client {
         const Environment: Environment;
     }
 
@@ -3171,4 +3147,4 @@ declare namespace OpenEO {
     export type ValidationResult = any;
 }
 
-export = OpenEO;
+export = Client;